+++ sktime/annotation/tests/test_all_annotators.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 import pytest
 from sktime.utils._testing.estimator_checks import _construct_instance, _make_args
-from sktime.utils import all_estimators
+from sktime.registry import all_estimators
 
 ALL_ANNOTATORS = all_estimators(estimator_types="series-annotator", return_names=False)
