+++ sktime/dists_kernels/tests/test_all_dist_kernels.py
@@ -2,7 +2,7 @@
 import numpy as np
 import pytest
 
-from sktime.utils import all_estimators
+from sktime.registry import all_estimators
 from sktime.utils._testing.panel import make_transformer_problem
 from sktime.tests._config import ESTIMATOR_TEST_PARAMS
