+++ sktime/forecasting/tests/test_sktime_forecasters.py
@@ -24,7 +24,7 @@ import pytest
 from sktime.forecasting.base import BaseForecaster
 from sktime.forecasting.base._sktime import _BaseWindowForecaster
 from sktime.forecasting.model_selection import temporal_train_test_split
-from sktime.utils import all_estimators
+from sktime.registry import all_estimators
 from sktime.utils._testing.estimator_checks import _construct_instance
 from sktime.utils._testing.forecasting import make_forecasting_problem
